\name{GFMaps-package}
\alias{GFMaps-package}
\alias{GFMaps}
\docType{package}
\title{
Visualization technique for interpretation of high-throughput genetic or proteomic experiments.
}
\description{
GFMaps integrates datasets derived from gene expression profiles with preexisting information from public
databases such as KEGG pathway or Gene Ontology using Gene Set Enrichment Analysis (GSEA) and introduce a framework that allow
us to visualize biologically annotated datasets and relevant ratings of genes via a color gradient.
}
\details{
\tabular{ll}{
Package: \tab GFMaps\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-01-29\cr
License: \tab GPL (version 2 or later)\cr
LazyLoad: \tab yes\cr
}
}

\author{
Dr. Sanjay Bhatikar \email{sanjay.bhatikar@biobase.in}, Kanika Arora \email{kanika.arora@biobase.in}

Maintainer: Kanika Arora \email{kanika.arora@biobase.in}
}
\references{ 
\itemize{
\item{1.}{ Subramanian, A., Tamayo, P., Mootha, V., K. Mukherjee, S., Ebert, B., L., Gillette, M., A., Paulvich, A., Pomeroy, S., L., Golub, T., R., Lander, E., S., Mesirov, J., P. (2005),Gene set enrichment analysis: A knowledge-based approach for interpreting genome-wide expression profiles \strong{PNAS 102, 15545-15550}}
\item{2.}{ Subramanian, A., Keuhn, H., Gould, J., Tamayo, P., Mesirov, J., P. (2007),GSEA-P: a desktop application for Gene Set Enrichment Analysis \strong{Bioinformatics 23(23), 3251-3253}}
 }
 }
\keyword{ package }
